Determine padding using the string length in bytes
Previously, the padding byte count was calculated based upon the number
of characters in the string.  This meant that on different platforms
with different encodings, the number of padding bytes varied causing
errors such as:

  SKIPPING DataType=0 for variable foo

Now the padding is always based upon the number of bytes in the string,
which should be fixed across platforms/encodings.